+++ app/assets/javascripts/actions/SearchActions.js
