+++ app/assets/javascripts/stores/SearchStore.js
